fix yaml bug
