+++ pkg/utils/prompt.go
@@ -26,27 +26,3 @@ func Confirm(question string) (bool, error) {
 func CanPrompt() bool {
 	return isatty.IsTerminal(os.Stderr.Fd())
 }
-
-func PickSlug(def string) (string, error) {
-	var slug string
-	if err := survey.AskOne(
-		&survey.Input{
-			Message: "Pick a unique identifier (slug) for this task",
-			Default: def,
-		},
-		&slug,
-		survey.WithStdio(os.Stdin, os.Stderr, os.Stderr),
-		// TODO: add a validator to ensure this slug is unique.
-		survey.WithValidator(func(val interface{}) error {
-			if str, ok := val.(string); !ok || !IsSlug(str) {
-				return errors.New("Slugs can only contain lowercase letters, underscores, and numbers.")
-			}
-
-			return nil
-		}),
-	); err != nil {
-		return "", errors.Wrap(err, "prompting for slug")
-	}
-
-	return slug, nil
-}
