+++ pkg/utils/slugs.go
@@ -1,9 +1,12 @@
 package utils
 
 import (
+	"os"
 	"strings"
 
+	"github.com/AlecAivazis/survey/v2"
 	"github.com/gosimple/slug"
+	"github.com/pkg/errors"
 )
 
 func init() {
@@ -25,4 +28,28 @@ func MakeSlug(s string) string {
 func IsSlug(text string) bool {
 	// The slug library will accept text with `-`'s, so we need to add our own check.
 	return slug.IsSlug(text) && !strings.Contains(text, "-")
+}
+
+func PickSlug(def string) (string, error) {
+	var slug string
+	if err := survey.AskOne(
+		&survey.Input{
+			Message: "Pick a unique identifier (slug) for this task:",
+			Default: def,
+		},
+		&slug,
+		survey.WithStdio(os.Stdin, os.Stderr, os.Stderr),
+		// TODO: add a validator to ensure this slug is unique.
+		survey.WithValidator(func(val interface{}) error {
+			if str, ok := val.(string); !ok || !IsSlug(str) {
+				return errors.New("Slugs can only contain lowercase letters, underscores, and numbers.")
+			}
+
+			return nil
+		}),
+	); err != nil {
+		return "", errors.Wrap(err, "prompting for slug")
+	}
+
+	return slug, nil
 }
